requirements.txt
ultralytics
confluent-kafka
PyYAML
dotenv-python
supervision
pymilvus
